Cleaned up and refactored some things for clarity
