+++ DatabaseTest/DatabaseTest.vcxproj.filters
@@ -21,9 +21,18 @@
     <ClCompile Include="string_format.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="sql_handler.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="string_format.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="sql_handler.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="unit_tests.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
