+++ DatabaseTest/string_format.cpp
@@ -75,3 +75,47 @@ std::vector<std::string> StringFormat::Tokenize(const std::string& str, const st
 
 	return tokens;
 }
+
+void StringFormat::CombineTokensInQuotes(std::vector<std::string>& tokens) {
+	bool openQuote = false;
+	size_t quoteStart = SIZE_MAX;
+
+	for (size_t i = 0; i < tokens.size(); i++) {
+		size_t quoteInd = tokens[i].find('"');
+
+		if (quoteInd != std::string::npos) {
+			// If there is a quotation mark remove it from the token
+			tokens[i].erase(quoteInd, 1);
+
+			// Then either close or open the quote
+			if (openQuote) {
+				// If the quote was closed, add this token to the starting one
+				tokens[quoteStart] += " " + tokens[i];
+				tokens.erase(tokens.begin() + i--);
+
+				openQuote = false;
+			}
+			else {
+				openQuote = true;
+			}
+
+			
+
+			// Handle cases where there is more than one mark in a token
+			quoteInd = tokens[i].find('"');
+			while (quoteInd != std::string::npos) {
+				openQuote = !openQuote;
+				tokens[i].erase(quoteInd, 1);
+				quoteInd = tokens[i].find('"');
+			}
+
+			if (openQuote)
+				quoteStart = i;
+		}
+		else {
+			// If there's still an open quote and that was unclosed, add this token to that one
+			tokens[quoteStart] += " " + tokens[i];
+			tokens.erase(tokens.begin() + i--);
+		}
+	}
+}
