+++ DatabaseTest/string_format.h
@@ -8,4 +8,5 @@ public:
 	static std::string AllCaps(const std::string& str);
 	static std::vector<std::string> Tokenize(const std::string& str, char delimiter);
 	static std::vector<std::string> Tokenize(const std::string& str, const std::string& delimiter);
+	static void CombineTokensInQuotes(std::vector<std::string>& tokens);
 };
