+++ DatabaseTest/unit_tests.h
@@ -1,0 +1,8 @@
+//#pragma once
+//#include "sql_handler.h"
+//#include "string_format.h"
+//
+//class UnitTests {
+//public:
+//	
+//};
